added new global flag USE_COUCH if on, cpymad will look at couch db for model definitions
currently doesn't work properly with multiprocessing,
so it is set to off.

restructured the cpymad model definitions somewhat:
flags -> knobs
a knob consist of list with name of variables and corresponding
unit value.
if a knob is "on", it means every variable will be set to one unit.
off should be that all variables are set to zero. makes sense?