Derive CpymadTest from object, instead of unittest.TestCase
For the unspecialized base class `CpymadTest`, it makes no sense to
execute the `test_xxx` methods defined herein. If derived from
`unittest.TestCase`, however, the `unittest` module can not distinguish
this class from a real test class and executes the `test_xxx` methods
anyway. This makes it impossible to execute the tests with a command
like:

    unittest -m test.cpymadtest.test_lhc