Add dependencies to numpy and Cython in setup.py
The `install_requires` dependency is important to make sure the target
system fulfills certain requirements independently from whether the
development (and build) system does - even when using binary
distribution.

NOTE: The `setup_requires` dependencies on numpy and Cython are kind of
pointless at the moment, since the modules are imported in the script
beforehand. Still, it seems like good practice to mention those.