Remove typechecks in cpymad.model.save_model
The python idiom of choice in most cases is duck typing.

Typechecks should mostly be used to differentiate between alternating
implementations for different use cases. And even then inheritance is
often the preferred mechanism.

I believe that this function made an instance method of the Model class
or better be removed from the module altogether:

- it is trivial
- this module shouldn't need to know about how models are stored on disk