Fix weird import bug
The last commit has created a weird import error, but only on python2.x
and only when installing with `python setup.py install` (as opposed to
`pip install .`).

Currently, I don't see any reason for this issue, especially since the
last commit was only a rename - I double checked that it was working on
the previous version.

Luckily, the issue is easy to fix.

In more detail: when performing `import cern.cpymad` an exception
occured with the following traceback:

    Traceback (most recent call last):
      File "test/cpymadtest/test_survey.py", line 2, in <module>
        from cern import cpymad
      File "build/bdist.linux-x86_64/egg/cern/cpymad/__init__.py", line 20, in <module>

      File "build/bdist.linux-x86_64/egg/cern/cpymad/model.py", line 35, in <module>
      File "build/bdist.linux-x86_64/egg/cern/madx.py", line 10, in <module>
      File "build/bdist.linux-x86_64/egg/cern/cpymad/madx.py", line 31, in <module>
    AttributeError: 'module' object has no attribute 'cpymad'