Enable sharing connections to remote process
The Madx object shouldn't assume to be the sole owner of the
LibMadxClient connection. I will add a Sequence type soon, that also
needs to talk to the remote process.

The LibMadxClient object will automatically close the connection when it
is garbage collected. Since there are no cyclic references in the RPC
code (I believe), the finalization will occur at the same time that the
last object using the connection is finalized, which is optimal.

Closing the connection implicitly has one further upside: we can pass
any libmadx object in the constructor without assuming that there is a
connection object supporting it. This means, if you feel that talking to
a remote process is too much overhead, you can do the following:

    import libmadx
    m = Madx(libmadx=libmadx)

which will make the Madx object directly use the libmadx module.