Use custom replacement for table_get_column
As explained in the comment MAD-X r4777 fixes the handling of integer
columns in the table_get_column function. As it seems overkill to
require a new MAD-X version for a minor patch, the function is replaced
with a Cython snippet inside cpymad for now.

When we require a newer version of MAD-X sometime in the future, THIS
COMMIT MAY BE REVERTED.

There is one additional benefit to the replacement code though: The
ability to raise exceptions with more detailed descriptions on what went
wrong.