Remove recursive_history feature
I don't believe the feature was very useful. It just added a
hard-to-understand parameter to the constructor of the Madx/Model
classes.

More importantly, the recursive_history mode, made it impossible to
distinguish between commands created by PyMAD and commands contained in
a CALLed file.

Furthermore, the implementation was probably still incorrect for many
edge cases like multiline commands.