Remove build time dependency on numpy
This makes the created binary specific to only the python build. It can
be used on target machines with different numpy versions than on the
system which was used to create the binary. This significantly
simplifies distribution of binaries for windows systems.

Furthermore, the setup.py script becomes a little easier, since we don't
require numpy to be installed in advance anymore.