Enforce lower-case keys for elements/summary
NOTE: the case for the table column keys cannot be normalized similarly,
as we need to pass the keys in its original case to the get_table_column
function.

For now, we just rely on MAD-X storing everything in lower-case
internally. If in the future, this turns out to be invalid, I see two
possible solutions:

- avoid proxy objects altogether (meaning get_table_column call time
  can always be made coinciding with the key normalization time)

- use a mapping [lower case -> original case] as is done in pydicti_ to
  recover the correct key case

.. _pydicti: https://pypi.python.org/pypi/pydicti/