Remove unused features: couch + ModelLocator
The ChainModelLocator is rather useless in real applications. These will
either use only a single source or use a more sophisticated mechanism.

The resource.couch module is unmaintained and untested. It should be
added later, when needed.

The DistinctModelLocator was developed for loading models from couchdb
and is untested and unmaintained.