Move some documentation to better locations
The very small examples do not justify separate documentation pages. By
moving them inside the docstrings they will be visible on the sphinx
pages as well, but also for anyone just browsing the source code or
using the python inline help() function.

The detailed rationale and description of the RPC module is best kept
exactly there… in the _rpc module.

Description of Madx / Model classes is best put in the docstring of
these classes.