Add getter functions for global variables to madx
The .lib file generated by `cmake -DCMAKE_GNUtoMS=ON` does not
contain exports of global variables. This commit right here is
an attempt to make cpymad compatible with a MAD-X version that
only exports functions.

Note that the two files in `madx/` need to be copied to `MAD-X/src`
before building MAD-X.