Distinguish between command parameters and other attributes
Store the parameters of the MAD-X command (i.e. the data content of
`struct command_parameter`) separately from other information that we
want to save with the element (e.g. members of the node/command/element
structs).

**The main benefit** is that this will allow using a more informative
datatype for the command parameters while keeping the other attributes
simple values.

**Note that**, from now on only command parameters will be visible when
iterating over the command/element contents and __getitem__ access is
now only possible for command parameters!

This makes it much easier to rigorously check the existence of command
parameters while constructing MAD-X statements for string input
(previously `util.mad_command` allowed passing names of other
attributes). AND: it will make it a lot less likely that another tool
accidentally includes the name of such an attribute in the parameters to
the command (by inspection).