Quote string arguments without escapes
`repr` is clearly not the right tool because it generates python
specific formatings (such as u'' and escaping of special characters).
MAD-X on the other hand doesn't do any unescaping, so we have to (!)
pass the untampered-with string.

Also, with this we quote strings with double-quotes by default which
looks much nicer! :)