Fix build for free-threaded python by running cython in each build
This removes the .c file from the source distribution and instead relies
on build configuration to provide a suitable cython version for each
build.

The motivation for this change is that cython>=3.1.0a1 is required for
supporting free-threaded python (3.13t). This fixes the following compile error
that we would get on free-threaded python by using a .c file generated with an
older cython version:

    unknown type name '__pyx_vectorcallfunc'; did you mean 'vectorcallfunc'

However, the .c file produced by cython3.1.0a1 only supports python>=3.8.

Hence, in order to support both py<3.8 and py>3.13, we need to compile
different .c files. Therefore, it makes to sense to include a .c file in a
common source distribution.